+++ common/src/main/java/de/crafty/eiv/common/config/instances/ClientConfig.java
@@ -8,6 +8,7 @@ public class ClientConfig extends AbstractEivConfig {
     private boolean background = true;
     private boolean itemWrapMode = true;
     private boolean appendModNamespace = true;
+    private boolean rightIndex = true;
 
     public ClientConfig() {
         super("client_settings");
@@ -37,11 +38,20 @@ public class ClientConfig extends AbstractEivConfig {
         this.appendModNamespace = appendModNamespace;
     }
 
+    public boolean isRightIndex() {
+        return rightIndex;
+    }
+
+    public void setRightIndex(boolean rightIndex) {
+        this.rightIndex = rightIndex;
+    }
+
     @Override
     protected void loadData() {
         this.background = this.data().get("background").getAsBoolean();
         this.itemWrapMode = this.data().get("itemWrapMode").getAsBoolean();
         this.appendModNamespace = this.data().get("appendModNamespace").getAsBoolean();
+        this.rightIndex = this.data().get("rightIndex").getAsBoolean();
     }
 
     @Override
@@ -49,5 +59,6 @@ public class ClientConfig extends AbstractEivConfig {
         this.data().addProperty("background", this.background);
         this.data().addProperty("itemWrapMode", this.itemWrapMode);
         this.data().addProperty("appendModNamespace", this.appendModNamespace);
+        this.data().addProperty("rightIndex", this.rightIndex);
     }
 }